booking-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
requests
psycopg2-binary
pymysql
cryptography
pika
